Completely reworked logic and representation
Day 6a solution had a serious flaw: It grew exponentially slower each day. Not sure how to remedy.
In this code now, we just forget about the order of fish, and do not inspect and update each one separately in order.
Instead, we just remember how many fish with each timer we originally saw. This is now codified as a class instead of a list.
Then we can quickly know how the distribution will be the next day, how many new fish will spawn, and how many fish there are in total.